Swap FileHelper methods for new Ini parser
