--- tests/fixtures/cms/filehelper/sections.ini
@@ -1,10 +1,0 @@
-var1 = "value 1"
-var2 = "value 21"
-
-[section]
-sectionVar1 = "section value 1"
-sectionVar2 = "section value 2"
-
-[section data]
-sectionVar3 = "section value 3"
-sectionVar4 = "section value 4"
