--- tests/fixtures/cms/filehelper/simple.ini
@@ -1,2 +1,0 @@
-var1 = "value 1"
-var2 = "value 21"
