--- tests/fixtures/cms/filehelper/subsections.ini
@@ -1,15 +1,0 @@
-var1 = "value 1"
-var2 = "value 21"
-
-[section]
-sectionVar1 = "section value 1"
-sectionVar2 = "section value 2"
-subsection[] = "subsection value 1"
-subsection[] = "subsection value 2"
-sectionVar3 = "section value 3"
-
-[section data]
-sectionVar3 = "section value 3"
-sectionVar4 = "section value 4"
-subsection[] = "subsection value 1"
-subsection[] = "subsection value 2"
